Frick you optifine
Had to change shared models into individual models for lightsabers, axe-olotls, dev totems and trapdoor shields. Effectively sacrificing reload performance on all platforms and my own future sanity when wanting to make large edits solely so that it functions correctly on the most commonly used mod for CIT features.

While I entirely see the need for a mod that has it all and essentially functions as a client like optifine, the fact that it is unable to support its own features to more than a basic extent makes it difficult for a creator dependant on it to keep projects properly updated. Alternates offer a trade off of better support for people willing to mess with an insane amount of mods on their client, they often add new technologies that I want to implement, but often cannot, due to the popularity of optifine. Development of a mod is no small task, keeping such a large mod updated at all is an incredible feat, but keeping the dev team limited to one member is becoming detrimental. I spent the last few hours copying and pasting unnecessarily because of it. Having more than one person supporting a mod like that and adding features is what will keep it from becoming a hindrance to the larger community. I know for a fact that people would be willing to join a dev team for optifine.